requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.1.0
Babel==2.15.0
beautifulsoup4==4.13.3
bleach==6.1.0
blinker==1.9.0
blis==1.2.0
bm25s==0.2.7.post1
brotlipy==0.7.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.2
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.4.2
docx2txt==0.8
duckduckgo_search==7.3.2
einops==0.8.1
expecttest==0.1.6
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.9.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.28.1
hypothesis==6.87.1
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.8
langchain-core==0.3.35
langcodes==3.5.0
langgraph==0.2.72
langgraph-checkpoint==2.0.13
langgraph-sdk==0.1.51
langsmith==0.3.8
language_data==1.3.0
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index==0.12.16
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-huggingface==0.5.1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-postprocessor-rankgpt-rerank==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.2
llama-parse==0.6.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.26.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook_shim==0.2.4
openai==1.61.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
platformdirs==4.2.2
preshed==3.0.9
primp==0.12.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.1
pypdf==5.3.0
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-etcd==0.4.5
python-json-logger==2.0.7
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.11.6
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.19.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
sentence-transformers==3.4.1
shellingham==1.5.4
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
streamlit==1.42.0
striprtf==0.0.26
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
torch==2.1.0
torchaudio==2.1.0
torchelastic==0.2.2
torchvision==0.16.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
triton==2.1.0
typer==0.15.1
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
virtualenv==20.26.3
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
